Refactor password reset forms and API
